Move rules to package where they will be when moved to litmus
