+++ nexus/nexus-test/nexus-testsuite-support/src/main/java/org/sonatype/sisu/litmus/testsupport/TestData.java
@@ -10,7 +10,7 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.testsuite.support;
+package org.sonatype.sisu.litmus.testsupport;
 
 import java.io.File;
